pyproject.toml
[project]
name = "sre-rca-agent"
version = "0.1.0"
description = "Vendor-neutral SRE investigation and root cause analysis (RCA) agent with schema-backed onboarding and provider adapters."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
keywords = ["sre", "rca", "incident", "investigation", "langgraph", "fastapi", "copilotkit"]
urls = { "Repository" = "https://github.com/rogerchucker/sre-rca-agent" }
dependencies = [
  "fastapi>=0.111",
  "uvicorn>=0.30",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "python-dotenv>=1.0",
  "pyyaml>=6.0",
  "httpx>=0.27",
  "langgraph>=0.2.0",
  "openai>=1.40.0",
  "jinja2>=3.1.6",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.2",
  "psycopg2>=2.9.11",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
minversion = "8.2"
testpaths = ["tests"]
addopts = "-q"
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
]
